+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.33.0] - 2019-10-02
 ### Added
 - Loading more results of a search result now changes the url. The back and forward feature is working properly to consider the page the user was last in.
 - Added a `show previous` button for when the user is not seeing the search results from the first item.